requirements.txt
pytest>=7.0
python-dotenv>=1.0
Flask>=2.0
cryptography>=45.0
requests>=2.32